backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pandas==2.3.2
python-docx==1.2.0
python-pptx==1.0.2
google-cloud-storage==3.3.1
google-cloud-firestore==2.21.0
google-cloud-secret-manager==2.24.0
azure-identity==1.24.0
azure-storage-blob==12.26.0
boto3==1.40.22
oci==2.159.0
requests==2.32.5
openpyxl==3.1.5
schedule==1.2.0
google-cloud-billing==1.12.0
